requirements.txt
Django==3.2.4
djangorestframework==3.12.4
requests==2.25.1



